Move disjunct and connector printing to print-dict.c
